Convert EOL characters to '\n'
Since we need this to be able to save to MongoDB, which does not
accept \r\n nor \n  we need to set the EOL/New line characters as
literal '\n' strings. And when parsing the string from MongoDB to
HTML, we will substitute the '\n' for html tag '<br>'.

So now when compiling the html it will become something like this:
"Hi\nI am a multilne\ncompiled markdown".